[Refactor] Service 계층 수정 - #62
