+++ src/main/java/com/appcenter/marketplace/domain/coupon/repository/CouponRepositoryCustom.java
@@ -8,7 +8,4 @@ import java.util.List;
 public interface CouponRepositoryCustom {
     List<CouponResDto> findOwnerCouponResDtoByMarketId(Long marketId);
     List<CouponMemberResDto> findMemberCouponResDtoByMarketId(Long marketId);
-    List<CouponLatestTopResDto> findLatestTopCouponDtoListByMarket(Integer size);
-    List<CouponMarketResDto> findLatestCouponMarketResDtoListByMarket(LocalDateTime modifiedAt, Long couponId, Integer size);
-    List<CouponClosingTopResDto> findClosingTopCouponDtoList(Integer size);
 }
